VideoEditor: replace `MediaActions.edit` with flux adapter calls
